Improve image handling and raster grid UI
Refactor PluginWindow UI and texture handling: finish texture compilation after loading previews, create and configure FSlateDynamicImageBrush via GetMyBrush, force volatile border for preview, wrap raster grid in SScaleBox and call Invalidate on RasterGridPanel for layout/volatility updates. Standardize loop index types and expand RasterCount from UINT8 to int to avoid overflow. Consolidate texture creation flow in OptimizeImage and apply UI-focused compression/never-stream/UpdateResource + FlushRenderingCommands. Minor cleanup (removed placeholder tab comment, added required includes).

+++ Plugins/GIS_Terrain_Generator/Source/GIS_Terrain_Generator/Public/PluginWindow.h
@@ -13,6 +13,7 @@
 #include "Widgets/Layout/SUniformGridPanel.h"
 #include "Widgets/Layout/SWidgetSwitcher.h"
 #include "Widgets/Layout/SScrollBox.h"
+#include "Widgets/Layout/SScaleBox.h"
 
 #include "Widgets/Text/STextBlock.h"
 #include "Widgets/Input/SButton.h"
@@ -30,6 +31,10 @@
 #include "Styling/StyleColors.h"
 
 #include "Brushes/SlateImageBrush.h"
+
+
+#include "Components/InvalidationBox.h"
+#include "Widgets/SInvalidationPanel.h"
 
 /**
  *
@@ -91,7 +96,7 @@ private:
 	bool PollRasterGeneration();
 	bool RasterDone{ false };
 
-	UINT8 RasterCount{ 0 };
+	int RasterCount{ 0 };
 	bool RasterCountDone{ false };
 
 	float RasterScaleX{ 64.0f };
